pyproject.toml
[project]
name = "hemoglobin-case-study"
version = "0.1.0"
description = "Scripts and docs for the hemoglobin beta E6V workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "biopython==1.86",
  "matplotlib==3.10.8",
  "numpy==2.4.3"
]

